backend/requirements.txt
supabase
fastapi
uvicorn
python-multipart
python-dotenv
reportlab
# Legacy (to be removed)
sqlalchemy
psycopg2-binary
gunicorn
